feat: migrate roles functions to own module
